♻️ refactor(tui): move the Commits view into views/commits.rs
First of the two pilot views (#635). `state/commits.rs`, the nine commit
renderers in `ui.rs` and the seven `App` methods in `app.rs` become one
module holding the state, the rendering and the key handling of one view.

The run loop's `match app.resolve_modal(KeyContext::Commits, key)` becomes
`App::handle_commits_key`, the shape `handle_working_tree_key` already
had. That is the point of #613: the order the arms resolve in is a
contract, and a `match` inside the event loop cannot be driven from a
test, so nothing pinned it. The arms are the run loop's, verbatim and in
the same order — in particular there is no `modal_toggle_stroke` block
here, because `c` closes this overlay as a bound `CommitsClose`
alternative and not as a rebindable global toggle. Adding one would
change which stroke wins.

`ui.rs` now imports `recent_commits_lines` and `draw_commits` back out of
`views::commits`. That reads as an inversion and is not one: the sidebar
pane that calls it is itself a view that has not been extracted yet, so
the call is view-to-view with one end still sitting in the old file.

Four `ui.rs` helpers the moved renderer needs widen to `pub(crate)`
(`centered`, `overlay_worktree`, `scrollable_body_area`) and `App::task_tx`
does too, since the two `request_*` workers moved out of `app.rs`.
`views` itself is a private module like `ui`, so the crate's public
surface is unchanged: every `gwm::tui::*` path the tests use still
resolves through the same re-exports, and `tests/` is untouched.

No behaviour change. `cargo test`: 111 suites, 3526 passed, 0 failed.
`cargo clippy --all-targets -- -D warnings`: clean.

refs #635
